fix: Added ABC class to a Abstract models in db/models, added user verification to admin endpoint.

+++ src/endpoints/admin.py
@@ -1,17 +1,18 @@
-from flask import Blueprint, render_template, Response, request, jsonify
+from flask import Blueprint
 from http import HTTPStatus
-from sqlalchemy.exc import IntegrityError
 from src.db import User
-from src.decorators import validate_model_request, handle_db_exception, require_auth
+from src.decorators import (
+    validate_model_request,
+    handle_db_exception,
+    require_auth
+)
 from src.models import (
     UserRoleUpdateModel,
-    UserEmailModel,
-    UserRoleModel,
     GenericIdModel,
 )
-from src.constants import UserRolesEnum
-from src.exceptions import InvalidDataError
-from src.constants import ErrorsMsgEnum
+from src.constants import UserRolesEnum, ErrorsMsgEnum
+from src.exceptions import AuthorizationError
+
 
 admin_route = Blueprint("admin_route", __name__, url_prefix="/admin")
 
@@ -21,9 +22,12 @@ admin_route = Blueprint("admin_route", __name__, url_prefix="/admin")
 @validate_model_request(UserRoleUpdateModel)
 @handle_db_exception(ErrorsMsgEnum.ERROR_USER_UPDATE_FAILED)
 def update_user_role(data: UserRoleUpdateModel):
-    # TODO: Add additional safety checks ( Check if user in question is not admin, etc)
-    User.update_by_id(data.id, **data.dict())
-    return "", HTTPStatus.NO_CONTENT
+    user_changed = User.get_by_id(data.id)
+    if user_changed.role == UserRolesEnum.ADMIN:
+        User.update_by_id(data.id, **data.dict())
+        return "", HTTPStatus.NO_CONTENT
+    else:
+        raise AuthorizationError(ErrorsMsgEnum.ERROR_ROLE_INVALID)
 
 
 @admin_route.route("/delete_user", methods=["PATCH"])
@@ -31,6 +35,9 @@ def update_user_role(data: UserRoleUpdateModel):
 @validate_model_request(GenericIdModel)
 @handle_db_exception(ErrorsMsgEnum.ERROR_DELETE_FAILED)
 def delete_user_by_id(data: GenericIdModel):
-    # TODO: Add additional safety checks  ( Check if user in question is not admin, etc)
-    User.delete_by_id(data.id)
-    return "", HTTPStatus.NO_CONTENT
+    user_changed = User.get_by_id(data.id)
+    if user_changed.role == UserRolesEnum.ADMIN:
+        User.delete_by_id(data.id)
+        return "", HTTPStatus.NO_CONTENT
+    else:
+        raise AuthorizationError(ErrorsMsgEnum.ERROR_ROLE_INVALID)
